Begins split of terms-list component in three.
